### What's New

- Project restructuring for modularity, see `Readme.md` in `src/`
- Data structure: added basic `vector` implementation (see [commit](https://code.harvard.edu/sam7740/cs165-2024-code/commit/6efd5e8971c8a5b28f0fd1a3cbef7bec6fd6a6c0) for more on its pros and cons) 
- bug fixes: use double for `avg` calculations for greater precision; correctly set the column datatype to `INT` on db startup and remember to free `current_db` at shutdown to avoid memory leaks. 

### Next Moves: 
- Still some memory leaks when compiled with sanitizers; needs to be fixed next. 
- Impl of other data structures like `b-trees` needed for milestone 2

+++ src/db/impl/query/handler.c
@@ -1,9 +1,9 @@
-#include "query_handler.h"
+#include "handler.h"
 
 #include <string.h>
 
 #include "client_context.h"
-#include "include/operators.h"
+#include "operators.h"
 #include "query_exec.h"
 #include "utils.h"
 
@@ -103,8 +103,8 @@ char *handle_print(DbOperator *query) {
       } else if (column->data_type == LONG) {
         long *data = (long *)column->data;
         printed = snprintf(current, remaining, "%ld", data[row]);
-      } else if (column->data_type == FLOAT) {
-        float *data = (float *)column->data;
+      } else if (column->data_type == DOUBLE) {
+        double *data = (double *)column->data;
         printed = snprintf(current, remaining, "%f", data[row]);
       } else {
         log_err("handle_print: Unsupported data type\n");
